+++ bfabric_scripts/src/bfabric_scripts/bfabric_save_workflowstep.py
@@ -19,6 +19,11 @@ Usage: bfabric_save_workflowstep.py 285507
 from __future__ import annotations
 
 import argparse
+import os
+from pathlib import Path
+
+import yaml
+from pydantic import BaseModel
 
 from bfabric import Bfabric
 from bfabric.entities import Workunit
@@ -32,31 +37,17 @@ def _get_user_id(login: str, client: Bfabric) -> int | None:
         raise RuntimeError(f"Could not find user with login: {login}")
 
 
-def save_workflowstep(workunit_id: int | None = None) -> None:
+class SaveWorkflowStepConfig(BaseModel):
+    template_step_ids: dict[int, int]
+    template_ids: dict[int, int]
+
+
+def save_workflowstep(workunit_id: int, config: SaveWorkflowStepConfig) -> None:
     """Creates an analysis workflow step for a given workunit id."""
     client = Bfabric.connect()
-    workflowtemplatestep_ids = {
-        224: 247,  # MaxQuant
-        # 295: 248, # FragPipe-RESOURCE
-        314: 254,  # DIANN
-        255: 256,  # maxquant_scaffold
-        266: 258,  # MaxQuant-sampleSizeEstimation
-        302: 259,
-        308: 259,
-        310: 259,
-        315: 290,
-    }
-    workflowtemplate_ids = {
-        224: 59,  # Proteomics Data analysis
-        # 295: 59,
-        314: 59,
-        255: 60,  # Proteomics Results
-        266: 60,
-        302: 60,
-        308: 60,
-        310: 60,
-        315: 60,
-    }
+
+    workflowtemplatestep_ids = config.template_step_ids
+    workflowtemplate_ids = config.template_ids
 
     workunit = Workunit.find(id=workunit_id, client=client)
     user_id = _get_user_id(login=workunit["createdby"], client=client)
@@ -101,8 +92,12 @@ def main() -> None:
     """Parses command line args and calls `save_workflowstep`."""
     parser = argparse.ArgumentParser(description="Create an analysis workflow step")
     parser.add_argument("workunitid", metavar="workunitid", type=int, help="workunit id")
+    parser.add_argument(
+        "config_path", default=Path(os.path.expanduser("~/slurmworker/config/legacy_template_steps.yml")), type=Path
+    )
     args = parser.parse_args()
-    save_workflowstep(workunit_id=args.workunitid)
+    config = SaveWorkflowStepConfig.model_validate(yaml.safe_load(args.config_path.read_text()))
+    save_workflowstep(workunit_id=args.workunitid, config=config)
 
 
 if __name__ == "__main__":